move testPopulateOtherFields test method to MigrationTestRpc
